+++ server/src/main/java/com/example/addition/server/service/AdditionService.java
@@ -1,9 +1,11 @@
 package com.example.addition.server.service;
 
+import java.math.BigInteger;
+
 import javax.servlet.http.HttpSession;
 
 import com.example.addition.server.exception.AdditionException;
 
 public interface AdditionService {
-	Long processData(String payloadRequest, HttpSession session) throws NumberFormatException, InterruptedException, AdditionException;	
+	BigInteger processData(String payloadRequest, HttpSession session) throws NumberFormatException, InterruptedException, AdditionException;	
 }
